[Refactor] Refactor the command container run.
